+++ src/matcha_ml/core/core.py
@@ -1,9 +1,15 @@
 """The core functions for matcha."""
 import os
-from typing import Optional
+from typing import List, Optional, Tuple
 
 from matcha_ml.cli._validation import get_command_validation
+from matcha_ml.cli.ui.print_messages import print_error, print_status
+from matcha_ml.cli.ui.status_message_builders import (
+    build_status,
+    build_step_success_status,
+)
 from matcha_ml.errors import MatchaError
+from matcha_ml.runners import AzureRunner
 from matcha_ml.services.global_parameters_service import GlobalParameters
 from matcha_ml.state import MatchaStateService, RemoteStateManager
 from matcha_ml.state.matcha_state import MatchaState
@@ -74,6 +80,41 @@ def get(
         )
 
     return result
+
+
+def destroy(resources: List[Tuple[str, str]]) -> None:
+    """Destroy resources.
+
+    Args:
+        resources (List[Tuple[str,str]): the list of resources to be actioned by the verb to be provided to the user as a status message
+
+    Raises:
+        Matcha Error: where no state has been provisioned.
+    """
+    remote_state_manager = RemoteStateManager()
+    if not remote_state_manager.is_state_provisioned():
+        print_error(
+            "Error - resources that have not been provisioned cannot be destroyed. Run 'matcha provision' to get started!"
+        )
+        raise MatchaError(
+            "Error - resources that have not been provisioned cannot be destroyed. Run 'matcha provision' to get started!"
+        )
+
+    with remote_state_manager.use_lock(), remote_state_manager.use_remote_state():
+        # create a runner for deprovisioning resource with Terraform service.
+        template_runner = AzureRunner()
+
+        if template_runner.is_approved(verb="destroy", resources=resources):
+            # deprovision the resources
+            template_runner.deprovision()
+            remote_state_manager.deprovision_remote_state()
+            print_status(build_step_success_status("Destroying resources is complete!"))
+        else:
+            print_status(
+                build_status(
+                    "You decided to cancel - your resources will remain active! If you change your mind, then run 'matcha destroy' again."
+                )
+            )
 
 
 def analytics_opt_out() -> None:
